Forenkling av delt bosted og endringsperioder ifm test
